requirements.txt
streamlit
sentence-transformers
faiss-cpu
scikit-learn
PyPDF2
nltk
